# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CMS Company API",
    "version": "1.0.0",
    "description": "Company resource backed by PostgreSQL."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "tags": [
    {
      "name": "Company",
      "description": "Singleton company record (first row in DB)"
    },
    {
      "name": "Employee",
      "description": "Employee utilities based on authenticated user"
    }
  ],
  "paths": {
    "/employee/id": {
      "get": {
        "tags": ["Employee"],
        "summary": "Get employeeId by bearer token",
        "operationId": "getEmployeeIdByToken",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Employee id for current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmployeeIdResponse" },
                "example": { "employeeId": "8a6f24ad-1f07-48d8-b4d4-24aa6da67ef8" }
              }
            }
          },
          "401": {
            "description": "Missing/invalid bearer token user id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Employee not found for current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/company": {
      "get": {
        "tags": ["Company"],
        "summary": "Get company",
        "operationId": "getCompany",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "EmployeeId",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Employee identifier from employee table."
          }
        ],
        "responses": {
          "200": {
            "description": "Company with employee count",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" },
                "example": {
                  "id": "company-1",
                  "name": "TrustFlow",
                  "employeeCount": 126
                }
              }
            }
          },
          "400": {
            "description": "Missing EmployeeId header",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Missing/invalid bearer token user id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "No company linked to current employee/user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Company"],
        "summary": "Update company name",
        "operationId": "patchCompany",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchCompanyBody" },
              "example": { "name": "TrustFlow" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated company",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Company not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Company"],
        "summary": "Delete company",
        "description": "Deletes the company and related employees (CASCADE).",
        "operationId": "deleteCompany",
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Company not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Company": {
        "type": "object",
        "required": ["id", "name", "employeeCount"],
        "properties": {
          "id": { "type": "string", "example": "company-1" },
          "name": { "type": "string", "example": "TrustFlow" },
          "employeeCount": { "type": "integer", "minimum": 0, "example": 126 }
        }
      },
      "PatchCompanyBody": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "description": "New company name (non-empty after trim)" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "EmployeeIdResponse": {
        "type": "object",
        "required": ["employeeId"],
        "properties": {
          "employeeId": { "type": "string" }
        }
      }
    }
  }
}